refactor lot and fix minor bugs
