+++ app/controllers/api_controller.rb
@@ -1,4 +1,9 @@
 # frozen_string_literal: true
 
 class ApiController < ApplicationController
+  private
+
+    def lot_not_found
+      render json: { error: "Lot did not found" }, status: :not_found
+    end
 end
