+++ app/controllers/bids_controller.rb
@@ -15,14 +15,14 @@ class BidsController < ApiController
         render json: bid.errors, status: :unprocessable_entity
       end
     else
-      render json: { error: "Current user can't create bid"  }, status: :forbidden
+      render json: { error: "Current user can't create bid" }, status: :forbidden
     end
   end
 
 
   def index
     bids = @lot.bids
-    render json: bids, each_serializer: BidSerializer,  current_user_id: current_user.id
+    render json: bids, each_serializer: BidSerializer, current_user_id: current_user.id
   end
 
   private
@@ -30,14 +30,14 @@ class BidsController < ApiController
     def check_lot_in_process
       @lot = Lot.find(bid_params[:lot_id])
       unless @lot && @lot.in_process?
-        render json: { error: "Lot did not found" }, status: :not_found
+        lot_not_found
       end
     end
 
     def check_lot_not_pending
       @lot = Lot.find(bid_params[:lot_id])
-      if !@lot && @lot.pending?
-        render json: { error: "Lot not found" }, status: :not_found
+      unless @lot && !@lot.pending?
+        lot_not_found
       end
     end
