+++ app/models/lot.rb
@@ -25,6 +25,7 @@
 #
 
 class Lot < ApplicationRecord
+  include LotJobsHelper
   mount_uploader :image, LotImageUploader
   belongs_to :user
   has_one :order, through: :bid
@@ -52,13 +53,6 @@ class Lot < ApplicationRecord
 
   after_create :create_jobs!
   before_update :check_start_and_end_time
-
-
-  def create_jobs!
-    to_in_process_job = create_status_job :in_process
-    to_close_job = create_status_job :closed
-    update_columns(lot_jid_in_process: to_in_process_job, lot_jid_closed: to_close_job)
-  end
 
   def check_start_and_end_time
     if lot_start_time_changed?
@@ -98,29 +92,9 @@ class Lot < ApplicationRecord
         .where(status: :pending)
   }
 
-
-  def create_status_job (status)
-    if status == :in_process
-      LotStatusUpdateWorker.perform_at(lot_start_time, id, status)
-    elsif status == :closed
-      LotStatusUpdateWorker.perform_at(lot_end_time, id, status)
-    end
-  end
-
-
-  def create_status_job! (status)
-    job = create_status_job status
-    if status == :in_process
-      update_column(:lot_jid_in_process, job)
-    elsif status == :closed
-      update_column(:lot_jid_closed, job)
-    end
-  end
-
   def check_estimated_prices(bid)
-    if (self.current_price >= self.estimated_price)
+    if (bid.proposed_price >= self.estimated_price)
       self.lot_jid_closed = nil
-      self.winning_bid = bid.id
       self.status = :closed
       save
     end
