+++ app/models/order.rb
@@ -20,6 +20,6 @@ class Order < ApplicationRecord
 
   enum status: { pending: 0, sent: 1, delivered: 2 }
   validates :arrival_location, :arrival_type, :status, presence: true
-  enum arrival_type: { pickup: 0, "Royal Mail".to_sym => 1, "United States Postal Service".to_sym => 2,
-                       "DHL Express".to_sym => 3 }
+  enum arrival_type: { pickup: 0, "Royal Mail" => 1, "United States Postal Service" => 2,
+                       "DHL Express" => 3 }
 end
