+++ spec/rubocop_spec.rb
@@ -2,6 +2,6 @@
 
 describe "Rubocop" do
   it "should execute without offenses" do
-    expect(system("rubocop")).to eq(true)
+    expect(system("rubocop --auto-correct")).to eq(true)
   end
 end
